Refactor C extension

+++ extensions/timer.h
@@ -1,5 +1,5 @@
 #include "Python.h"
 
-long long hpTimer(void);
+uint64_t hpTimer(void);
 double hpTimerUnit(void);
 extern const char HP_TIMER_IMPLEMENTATION[];
